requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-pptx>=0.6.21
Pillow>=10.0.0
requests>=2.31.0
certifi>=2024.0.0
